+++ components/cluster/command/upgrade.go
@@ -14,21 +14,7 @@
 package command
 
 import (
-	"errors"
-	"fmt"
-	"os"
-
-	"github.com/joomcode/errorx"
-	perrs "github.com/pingcap/errors"
-	"github.com/pingcap/tiup/pkg/cluster/clusterutil"
-	operator "github.com/pingcap/tiup/pkg/cluster/operation"
-	"github.com/pingcap/tiup/pkg/cluster/spec"
-	"github.com/pingcap/tiup/pkg/cluster/task"
-	"github.com/pingcap/tiup/pkg/logger/log"
-	"github.com/pingcap/tiup/pkg/meta"
-	"github.com/pingcap/tiup/pkg/version"
 	"github.com/spf13/cobra"
-	"golang.org/x/mod/semver"
 )
 
 func newUpgradeCmd() *cobra.Command {
@@ -44,7 +30,8 @@ func newUpgradeCmd() *cobra.Command {
 			version := args[1]
 			teleCommand = append(teleCommand, scrubClusterName(clusterName))
 			teleCommand = append(teleCommand, version)
-			return upgrade(clusterName, version, gOpt)
+
+			return deployer.Upgrade(clusterName, version, gOpt)
 		},
 	}
 	cmd.Flags().BoolVar(&gOpt.Force, "force", false, "Force upgrade without transferring PD leader")
@@ -52,166 +39,4 @@ func newUpgradeCmd() *cobra.Command {
 	cmd.Flags().BoolVarP(&gOpt.IgnoreConfigCheck, "ignore-config-check", "", false, "Ignore the config check result")
 
 	return cmd
-}
-
-func versionCompare(curVersion, newVersion string) error {
-	// Can always upgrade to 'nightly' event the current version is 'nightly'
-	if newVersion == version.NightlyVersion {
-		return nil
-	}
-
-	switch semver.Compare(curVersion, newVersion) {
-	case -1:
-		return nil
-	case 0, 1:
-		return perrs.Errorf("please specify a higher version than %s", curVersion)
-	default:
-		return perrs.Errorf("unreachable")
-	}
-}
-
-func upgrade(clusterName, clusterVersion string, opt operator.Options) error {
-	exist, err := tidbSpec.Exist(clusterName)
-	if err != nil {
-		return perrs.AddStack(err)
-	}
-
-	if !exist {
-		return perrs.Errorf("cannot upgrade non-exists cluster %s", clusterName)
-	}
-
-	metadata, err := spec.ClusterMetadata(clusterName)
-	if err != nil && !errors.Is(perrs.Cause(err), meta.ErrValidate) {
-		return err
-	}
-
-	var (
-		downloadCompTasks []task.Task // tasks which are used to download components
-		copyCompTasks     []task.Task // tasks which are used to copy components to remote host
-
-		uniqueComps = map[string]struct{}{}
-	)
-
-	if err := versionCompare(metadata.Version, clusterVersion); err != nil {
-		return err
-	}
-
-	hasImported := false
-	for _, comp := range metadata.Topology.ComponentsByUpdateOrder() {
-		for _, inst := range comp.Instances() {
-			version := spec.ComponentVersion(inst.ComponentName(), clusterVersion)
-			if version == "" {
-				return perrs.Errorf("unsupported component: %v", inst.ComponentName())
-			}
-			compInfo := componentInfo{
-				component: inst.ComponentName(),
-				version:   version,
-			}
-
-			// Download component from repository
-			key := fmt.Sprintf("%s-%s-%s-%s", compInfo.component, compInfo.version, inst.OS(), inst.Arch())
-			if _, found := uniqueComps[key]; !found {
-				uniqueComps[key] = struct{}{}
-				t := task.NewBuilder().
-					Download(inst.ComponentName(), inst.OS(), inst.Arch(), version).
-					Build()
-				downloadCompTasks = append(downloadCompTasks, t)
-			}
-
-			deployDir := clusterutil.Abs(metadata.User, inst.DeployDir())
-			// data dir would be empty for components which don't need it
-			dataDirs := clusterutil.MultiDirAbs(metadata.User, inst.DataDir())
-			// log dir will always be with values, but might not used by the component
-			logDir := clusterutil.Abs(metadata.User, inst.LogDir())
-
-			// Deploy component
-			tb := task.NewBuilder()
-			if inst.IsImported() {
-				switch inst.ComponentName() {
-				case spec.ComponentPrometheus, spec.ComponentGrafana, spec.ComponentAlertManager:
-					tb.CopyComponent(
-						inst.ComponentName(),
-						inst.OS(),
-						inst.Arch(),
-						version,
-						"", // use default srcPath
-						inst.GetHost(),
-						deployDir,
-					)
-				}
-				hasImported = true
-			}
-
-			// backup files of the old version
-			tb = tb.BackupComponent(inst.ComponentName(), metadata.Version, inst.GetHost(), deployDir)
-
-			// copy dependency component if needed
-			switch inst.ComponentName() {
-			case spec.ComponentTiSpark:
-				tb = tb.DeploySpark(inst, version, "" /* default srcPath */, deployDir)
-			default:
-				tb = tb.CopyComponent(
-					inst.ComponentName(),
-					inst.OS(),
-					inst.Arch(),
-					version,
-					"", // use default srcPath
-					inst.GetHost(),
-					deployDir,
-				)
-			}
-
-			tb.InitConfig(
-				clusterName,
-				clusterVersion,
-				inst,
-				metadata.User,
-				opt.IgnoreConfigCheck,
-				meta.DirPaths{
-					Deploy: deployDir,
-					Data:   dataDirs,
-					Log:    logDir,
-					Cache:  spec.ClusterPath(clusterName, spec.TempConfigPath),
-				},
-			)
-			copyCompTasks = append(copyCompTasks, tb.Build())
-		}
-	}
-
-	// handle dir scheme changes
-	if hasImported {
-		if err := spec.HandleImportPathMigration(clusterName); err != nil {
-			return err
-		}
-	}
-
-	t := task.NewBuilder().
-		SSHKeySet(
-			spec.ClusterPath(clusterName, "ssh", "id_rsa"),
-			spec.ClusterPath(clusterName, "ssh", "id_rsa.pub")).
-		ClusterSSH(metadata.Topology, metadata.User, gOpt.SSHTimeout).
-		Parallel(downloadCompTasks...).
-		Parallel(copyCompTasks...).
-		ClusterOperate(metadata.Topology, operator.UpgradeOperation, opt).
-		Build()
-
-	if err := t.Execute(task.NewContext()); err != nil {
-		if errorx.Cast(err) != nil {
-			// FIXME: Map possible task errors and give suggestions.
-			return err
-		}
-		return perrs.Trace(err)
-	}
-
-	metadata.Version = clusterVersion
-	if err := spec.SaveClusterMeta(clusterName, metadata); err != nil {
-		return perrs.Trace(err)
-	}
-	if err := os.RemoveAll(spec.ClusterPath(clusterName, "patch")); err != nil {
-		return perrs.Trace(err)
-	}
-
-	log.Infof("Upgraded cluster `%s` successfully", clusterName)
-
-	return nil
 }
