+++ components/dm/command/upgrade.go
@@ -13,24 +13,8 @@
 
 package command
 
-/*
 import (
-	"errors"
-	"fmt"
-	"os"
-
-	"github.com/joomcode/errorx"
-	perrs "github.com/pingcap/errors"
-	"github.com/pingcap/tiup/pkg/cluster/clusterutil"
-	"github.com/pingcap/tiup/pkg/cluster/meta"
-	operator "github.com/pingcap/tiup/pkg/cluster/operation"
-	"github.com/pingcap/tiup/pkg/cluster/task"
-	"github.com/pingcap/tiup/pkg/logger"
-	"github.com/pingcap/tiup/pkg/logger/log"
-	"github.com/pingcap/tiup/pkg/repository/v0manifest"
-	"github.com/pingcap/tiup/pkg/utils"
 	"github.com/spf13/cobra"
-	"golang.org/x/mod/semver"
 )
 
 func newUpgradeCmd() *cobra.Command {
@@ -42,8 +26,7 @@ func newUpgradeCmd() *cobra.Command {
 				return cmd.Help()
 			}
 
-			logger.EnableAuditLog()
-			return upgrade(args[0], args[1], gOpt)
+			return deployer.Upgrade(args[0], args[1], gOpt)
 		},
 	}
 	cmd.Flags().BoolVar(&gOpt.Force, "force", false, "Force upgrade won't transfer leader")
@@ -52,130 +35,3 @@ func newUpgradeCmd() *cobra.Command {
 
 	return cmd
 }
-
-func versionCompare(curVersion, newVersion string) error {
-	// Can always upgrade to 'nightly' event the current version is 'nightly'
-	if v0manifest.Version(newVersion).IsNightly() {
-		return nil
-	}
-
-	switch semver.Compare(curVersion, newVersion) {
-	case -1:
-		return nil
-	case 0, 1:
-		return perrs.Errorf("please specify a higher version than %s", curVersion)
-	default:
-		return perrs.Errorf("unreachable")
-	}
-}
-
-func upgrade(clusterName, clusterVersion string, opt operator.Options) error {
-	if utils.IsNotExist(meta.ClusterPath(clusterName, meta.MetaFileName)) {
-		return perrs.Errorf("cannot upgrade non-exists cluster %s", clusterName)
-	}
-
-	metadata, err := meta.DMMetadata(clusterName)
-	if err != nil && !errors.Is(perrs.Cause(err), meta.ValidateErr) {
-		return err
-	}
-
-	var (
-		downloadCompTasks []task.Task // tasks which are used to download components
-		copyCompTasks     []task.Task // tasks which are used to copy components to remote host
-
-		uniqueComps = map[string]struct{}{}
-	)
-
-	if err := versionCompare(metadata.Version, clusterVersion); err != nil {
-		return err
-	}
-
-	for _, comp := range metadata.Topology.ComponentsByStartOrder() {
-		for _, inst := range comp.Instances() {
-			version := meta.ComponentVersion(inst.ComponentName(), clusterVersion)
-			if version == "" {
-				return perrs.Errorf("unsupported component: %v", inst.ComponentName())
-			}
-			compInfo := componentInfo{
-				component: inst.ComponentName(),
-				version:   v0manifest.Version(version),
-			}
-
-			// Download component from repository
-			key := fmt.Sprintf("%s-%s-%s-%s", compInfo.component, compInfo.version, inst.OS(), inst.Arch())
-			if _, found := uniqueComps[key]; !found {
-				uniqueComps[key] = struct{}{}
-				t := task.NewBuilder().
-					Download(inst.ComponentName(), inst.OS(), inst.Arch(), version).
-					Build()
-				downloadCompTasks = append(downloadCompTasks, t)
-			}
-
-			deployDir := clusterutil.Abs(metadata.User, inst.DeployDir())
-			// data dir would be empty for components which don't need it
-			dataDirs := clusterutil.MultiDirAbs(metadata.User, inst.DataDir())
-			// log dir will always be with values, but might not used by the component
-			logDir := clusterutil.Abs(metadata.User, inst.LogDir())
-
-			// Deploy component
-			tb := task.NewBuilder()
-			if inst.IsImported() {
-				switch inst.ComponentName() {
-				case meta.ComponentPrometheus, meta.ComponentGrafana, meta.ComponentAlertManager:
-					tb.CopyComponent(inst.ComponentName(), inst.OS(), inst.Arch(), version, inst.GetHost(), deployDir)
-				default:
-					tb.BackupComponent(inst.ComponentName(), metadata.Version, inst.GetHost(), deployDir).
-						CopyComponent(inst.ComponentName(), inst.OS(), inst.Arch(), version, inst.GetHost(), deployDir)
-				}
-				tb.InitConfig(
-					clusterName,
-					clusterVersion,
-					inst,
-					metadata.User,
-					opt.IgnoreConfigCheck,
-					meta.DirPaths{
-						Deploy: deployDir,
-						Data:   dataDirs,
-						Log:    logDir,
-						Cache:  meta.ClusterPath(clusterName, meta.TempConfigPath),
-					},
-				)
-			} else {
-				tb.BackupComponent(inst.ComponentName(), metadata.Version, inst.GetHost(), deployDir).
-					CopyComponent(inst.ComponentName(), inst.OS(), inst.Arch(), version, inst.GetHost(), deployDir)
-			}
-			copyCompTasks = append(copyCompTasks, tb.Build())
-		}
-	}
-
-	t := task.NewBuilder().
-		SSHKeySet(
-			meta.ClusterPath(clusterName, "ssh", "id_rsa"),
-			meta.ClusterPath(clusterName, "ssh", "id_rsa.pub")).
-		ClusterSSH(metadata.Topology, metadata.User, gOpt.SSHTimeout).
-		Parallel(downloadCompTasks...).
-		Parallel(copyCompTasks...).
-		ClusterOperate(metadata.Topology, operator.UpgradeOperation, opt).
-		Build()
-
-	if err := t.Execute(task.NewContext()); err != nil {
-		if errorx.Cast(err) != nil {
-			// FIXME: Map possible task errors and give suggestions.
-			return err
-		}
-		return perrs.Trace(err)
-	}
-
-	metadata.Version = clusterVersion
-	if err := meta.SaveDMMeta(clusterName, metadata); err != nil {
-		return perrs.Trace(err)
-	}
-	if err := os.RemoveAll(meta.ClusterPath(clusterName, "patch")); err != nil {
-		return perrs.Trace(err)
-	}
-
-	log.Infof("Upgraded cluster `%s` successfully", clusterName)
-
-	return nil
-}
-*/
